Java, PHP applications and MySQL service.
Java and PHP entities were added which are based on CloudFoundryWebAppImpl.
These classes allow to deploy an application in CloudFoundry and bind a set of services.

ClearDB (MySQL) service entity was added, which is based on  CloudFoundryServiceImpl.
It allows instantiating  a ClearDB service in CloudFoundry.

Test: several LiveTest are added to check the behavior of the aforementioned entities.
WIP:  ASAP we added the connection establishment between an application and the real service.